pyproject.toml
[project]
name = "tps"
version = "0.1.0"
description = "Translation Proxy System - 高可用、低成本的翻譯中介層"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "deepl>=1.25.0",
    "fastapi>=0.123.9",
    "google-cloud-translate>=3.23.0",
    "openai>=2.9.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
]

[project.scripts]
tps = "tps.app:main"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.uv]
package = true
